microk8s working

+++ back/init.sh
@@ -4,22 +4,43 @@
 test ! -d "$HOME/.k8s-project" && mkdir -p "$HOME/.k8s-project"
 test ! -d "$HOME/.k8s-project/tools" && mkdir -p "$HOME/.k8s-project/tools"
 
-#Check if helm tool exists
-if test ! -f "$HOME/.k8s-project/tools/helm"; then
-    curl -o $HOME/.k8s-project/tools/helm.tar.gz "https://get.helm.sh/helm-v3.0.0-linux-amd64.tar.gz"
-    cd $HOME/.k8s-project/tools
-    tar zxf helm.tar.gz --strip-components=1 linux-amd64/helm
-    rm -rf *.tar.gz
-    chmod +x $HOME/.k8s-project/tools/helm
-fi
 
-#Check if kubectl exists
-if test ! -f "$HOME/.k8s-project/tools/kubectl"; then
-    curl -o $HOME/.k8s-project/tools/kubectl "https://storage.googleapis.com/kubernetes-release/release/v1.15.1/bin/linux/amd64/kubectl"
-    chmod +x $HOME/.k8s-project/tools/kubectl
-fi
+if [[ "$OSTYPE" == "linux-gnu" ]]; then
 
-export PATH=$PATH:$HOME/.k8s-project/tools/
+    #Check if helm tool exists
+    if test ! -f "$HOME/.k8s-project/tools/helm"; then
+        curl -o $HOME/.k8s-project/tools/helm.tar.gz "https://get.helm.sh/helm-v3.2.0-linux-amd64.tar.gz"
+        cd $HOME/.k8s-project/tools
+        tar zxf helm.tar.gz --strip-components=1 linux-amd64/helm
+        rm -rf *.tar.gz
+        chmod +x $HOME/.k8s-project/tools/helm
+    fi
+
+    #Check if kubectl exists
+    if test ! -f "$HOME/.k8s-project/tools/kubectl"; then
+        curl -o $HOME/.k8s-project/tools/kubectl "https://storage.googleapis.com/kubernetes-release/release/v1.18.2/bin/linux/amd64/kubectl"
+        chmod +x $HOME/.k8s-project/tools/kubectl
+    fi
+    export PATH=$PATH:$HOME/.k8s-project/tools/
+
+elif [[ "$OSTYPE" == "darwin"* ]]; then
+
+    HELM=$(brew list|grep helm|wc -l)
+    KUBERNETES=$(brew list|grep kubernetes-cli|wc -l)
+
+    if test `echo $HELM` -eq 0; then
+        brew install helm
+    fi
+
+    if test `echo $KUBERNETES` -eq 0; then
+        brew install kubernetes-cli
+    fi
+    
+elif [[ "$OSTYPE" == "win32" ]]; then
+    echo "soon...";
+else
+    exit 4
+fi
 
 #Check if .kube/config exists
 if test ! -f "$HOME/.kube/config"; then
@@ -34,28 +55,40 @@ fi
 
 kubectl create namespace k8s-project
 
-#Configure tiller
-kubectl --namespace kube-system create sa tiller
-kubectl create clusterrolebinding tiller --clusterrole cluster-admin --serviceaccount=kube-system:tiller
-
-#Install helm
-helm init tiller --output yaml | sed 's@apiVersion: extensions/v1beta1@apiVersion: apps/v1@' | sed 's@  replicas: 1@  replicas: 1\n  selector: {"matchLabels": {"app": "helm", "name": "tiller"}}@' | kubectl apply -f -
-
-#Install custom ingress
-helm install k8s-project-ingress stable/nginx-ingress --namespace k8s-project --set controller.ingressClass="k8s-project-ingress"
+#Install custom ingress if not microk8s
+INGRESS_MICRO_K8S=$(kubectl -n ingress get pods|grep nginx-ingress-microk8s-controller|wc -l)
+if test `echo $INGRESS_MICRO_K8S` -eq 0; then
+    helm repo add stable https://kubernetes-charts.storage.googleapis.com/
+    helm repo update
+    helm install k8s-project-ingress stable/nginx-ingress --namespace k8s-project --set controller.ingressClass="k8s-project-ingress"
+    INGRESS_IP=$(kubectl get services --namespace k8s-project|grep k8s-project-ingress-nginx-ingress-controller|awk '{print $4}')
+else
+    if [[ "$OSTYPE" == "darwin"* ]]; then
+        INGRESS_IP=$(multipass info microk8s-vm | grep IPv4 | awk '{ print $2 }') 
+    else
+        INGRESS_IP=127.0.0.1
+    fi
+    echo "Ingress already installed with microk8s"
+fi
 
 #store ingressIp
-INGRESS_IP=$(kubectl get services --namespace k8s-project|grep k8s-project-ingress-nginx-ingress-controller|awk '{print $4}')
 kubectl delete --namespace k8s-project configmap k8s-project-config
 kubectl create configmap k8s-project-config --namespace k8s-project --from-literal=INGRESS_IP=$INGRESS_IP
+
+#cert-manager for ingress
+#kubectl apply --validate=false -f https://github.com/jetstack/cert-manager/releases/download/v0.14.3/cert-manager.crds.yaml
+#helm repo add jetstack https://charts.jetstack.io
+#helm repo update
+#helm install k8s-project-cert-manager jetstack/cert-manager --namespace k8s-project
+#kubectl apply -f ./cert-manager.yml
 
 #Install postgresql
 POSTGRES_INSTALLED=$(helm ls --namespace k8s-project|grep k8s-project-postgresql|wc -l)
 if test $POSTGRES_INSTALLED -eq 0
 then
     echo "Installing postgresql"
-    helm install --wait stable/postgresql k8s-project-postgresql --namespace k8s-project --set postgresqlUsername=k8sproject --set postgresqlDatabase=k8sproject --set initdbScripts."init\.sql"="CREATE DATABASE keycloak;GRANT ALL PRIVILEGES ON DATABASE keycloak TO k8sproject;CREATE DATABASE gitbucket;GRANT ALL PRIVILEGES ON DATABASE gitbucket TO k8sproject;CREATE DATABASE wiki;GRANT ALL PRIVILEGES ON DATABASE wiki TO k8sproject;"
-    
+    helm repo add bitnami https://charts.bitnami.com/bitnami
+    helm install --wait k8s-project-postgresql bitnami/postgresql --namespace k8s-project --set postgresqlUsername=k8sproject --set postgresqlDatabase=k8sproject --set initdbScripts."init\.sql"="CREATE DATABASE keycloak;GRANT ALL PRIVILEGES ON DATABASE keycloak TO k8sproject;CREATE DATABASE gitbucket;GRANT ALL PRIVILEGES ON DATABASE gitbucket TO k8sproject;CREATE DATABASE wiki;GRANT ALL PRIVILEGES ON DATABASE wiki TO k8sproject;"
 fi
 
 KEYCLOAK_INSTALLED=$(helm ls --namespace k8s-project|grep k8s-project-keycloak|wc -l)
